monitoreogrid-master/requirements.txt
Flask~=1.1.2
Flask-Bootstrap
Flask-Login
Flask-SQLAlchemy
Flask-WTF

#tables
SQLAlchemy~=1.3.22
pydeck~=0.5.0
pandas~=1.1.5
cryptography
dictalchemy
plotly~=4.14.1
scipy~=1.6.0
Flask-Migrate
python-dotenv~=0.15.0
pymysql
gunicorn
gevent
scikit-learn~=0.24.0
matplotlib
dictalchemy

xgboost~=1.3.3
pytz~=2020.5
email_validator
geopy~=2.0.0
requests~=2.25.1
googlemaps~=2.5.1
statsmodels~=0.12.1
seaborn~=0.11.1
networkx~=2.5

numpy~=1.21
osmnx
mysql-connector-python~=8.0.23
geopandas~=0.8.1
sklearn~=0.0
wtforms~=2.3.2
werkzeug~=1.0.1
shapely~=1.7.1
alembic~=1.4.3